+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For details about compatibility between different releases, see the **Commitment
 
 ### Added
 
+- Allow setting an expiry date for API keys in the Console
 - New event type `gs.gateway.connection.stats` with connection statistics. A new event is sent at most every `gs.update-connection-stats-debounce-time` time.
 
 ### Changed